+++ src/io/ipc/convert.rs
@@ -191,7 +191,7 @@ fn get_data_type(field: ipc::Field, extension: Extension, may_be_dictionary: boo
         ipc::Type::LargeUtf8 => DataType::LargeUtf8,
         ipc::Type::FixedSizeBinary => {
             let fsb = field.type_as_fixed_size_binary().unwrap();
-            DataType::FixedSizeBinary(fsb.byteWidth())
+            DataType::FixedSizeBinary(fsb.byteWidth() as usize)
         }
         ipc::Type::FloatingPoint => {
             let float = field.type_as_floating_point().unwrap();
@@ -273,7 +273,7 @@ fn get_data_type(field: ipc::Field, extension: Extension, may_be_dictionary: boo
                 panic!("expect a list to have one child")
             }
             let fsl = field.type_as_fixed_size_list().unwrap();
-            DataType::FixedSizeList(Box::new(children.get(0).into()), fsl.listSize())
+            DataType::FixedSizeList(Box::new(children.get(0).into()), fsl.listSize() as usize)
         }
         ipc::Type::Struct_ => {
             let mut fields = vec![];
